Refactored PackageRetrieval hooks
